for Py3.11: Split build and test due to cyclic dep
These are just temporary changes and they will be overwritten by the next
re-rendering of the feedstock.
The cyclic dependency on mp-api prevents solving of the test environment
until a variant for the new Python version is built once. Splitting the
test off as a separate step lets us do exactly that in one CI run.